requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyJWT==2.1.0
Werkzeug==2.0.1
Flask-Limiter==2.4.0
python-dotenv==0.19.0
SQLAlchemy==1.4.23